requirements.txt
Package             Version
------------------- -------------------
altair              4.1.0
beautifulsoup4      4.9.3
click               7.1.2
emoji               1.2.0
expertai-nlapi      2.1.3
h11                 0.12.0
idna                2.10
importlib-metadata  3.3.0
itsdangerous        1.1.0
jaraco.classes      3.2.0
jaraco.collections  3.2.0
jaraco.functools    3.2.0
jaraco.text         3.5.0
jedi                0.17.2
Jinja2              2.11.2
joblib              1.0.0
jsonschema          3.2.0
kiwisolver          1.3.1
lxml                4.6.2
MarkupSafe          1.1.1
matplotlib          3.3.4
mistune             0.8.4
more-itertools      8.7.0
msgpack             1.0.2
multidict           5.1.0
nbclient            0.5.1
nbconvert           6.0.7
nbformat            5.0.8
numpy               1.19.5
packaging           20.8
pandas              1.2.1
pandocfilters       1.4.3
parso               0.7.1
Pattern             3.6
pexpect             4.8.0
pickleshare         0.7.5
Pillow              8.1.0
pip                 20.3.1
plotly              4.14.3
portend             2.7.1
prometheus-client   0.9.0
prompt-toolkit      3.0.8
protobuf            3.15.1
psutil              5.8.0
ptyprocess          0.6.0
pyarrow             3.0.0
pycparser           2.20
pydeck              0.6.1
pyee                7.0.4
Pygments            2.7.3
pyparsing           2.4.7
pyrsistent          0.17.3
python-dateutil     2.8.1
python-docx         0.8.10
python-dotenv       0.15.0
pytz                2020.5
pyzmq               20.0.0
qtconsole           5.0.1
QtPy                1.9.0
regex               2020.11.13
requests            2.25.1
retrying            1.3.3
scipy               1.6.0
Send2Trash          1.5.0
setuptools          51.0.0.post20201207
sgmllib3k           1.0.0
six                 1.15.0
smmap               3.0.5
sortedcontainers    2.3.0
soupsieve           2.2
streamlit           0.76.0
tqdm                4.57.0
tweet-preprocessor  0.6.0
typing-extensions   3.7.4.3
urllib3             1.26.2
wordcloud           1.8.1
